pipeline/v0/requirements.txt
psycopg2-binary==2.9.7
apache-airflow==2.7.0